Deployment/requirements.txt
torch
torchvision
torchaudio
numpy
pillow
flask
python-dotenv
Werkzeug
